fix: removed code duplication
